# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agenda+ API",
        "description": "API RESTful para sistema de gestão de consultas médicas Agenda+",
        "contact": {
            "name": "Suporte Agenda+",
            "email": "suporte@agendaplus.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor da API"
        }
    ],
    "paths": {
        "/admin/doctors": {
            "get": {
                "tags": [
                    "Administração"
                ],
                "summary": "Listar médicos",
                "description": "Lista todos os médicos cadastrados. Apenas administradores podem acessar.",
                "operationId": "caf8a50c03ad30b07043171a326ff742",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de médicos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Apenas administradores"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Administração"
                ],
                "summary": "Cadastrar médico",
                "description": "Cria um novo médico. CRM deve ser único. Senha mínima de 8 caracteres.",
                "operationId": "7c2433806210dc3db74278a9ed116ca7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "crm",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Dr. João Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Mínimo 8 caracteres",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "crm": {
                                        "type": "string",
                                        "example": "123456-SP"
                                    },
                                    "specialty": {
                                        "type": "string",
                                        "example": "Cardiologia"
                                    },
                                    "health_insurances": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Médico cadastrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados inválidos ou CRM duplicado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/doctors/{id}": {
            "get": {
                "tags": [
                    "Administração"
                ],
                "summary": "Obter detalhes do médico",
                "description": "Retorna os detalhes completos de um médico",
                "operationId": "d36ac29baead9f862c3d4e7f0c3ccfb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do médico",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Médico não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Administração"
                ],
                "summary": "Atualizar médico",
                "description": "Atualiza os dados de um médico",
                "operationId": "e140dcc384e108e5c84311e822e10a8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Médico atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Administração"
                ],
                "summary": "Inativar médico",
                "description": "Inativa um médico usando soft delete",
                "operationId": "05ce810a3790daf7992040719c22234b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Médico inativado"
                    },
                    "404": {
                        "description": "Médico não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticação Sanctum",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}